feat: move logic to app.init()
